Cleanup and move variables and functions to data stores
